Rename demo application to v24-spring
