+++ appsec-kit-demo-v24-spring/frontend/themes/appsec-kit-demo-v24-spring/theme.json
