+++ appsec-kit-demo-v24-spring/mvnw.cmd
