+++ appsec-kit-demo-v24-spring/src/main/resources/META-INF/resources/icons/icon.png
