+++ appsec-kit-demo-v24-spring/src/main/resources/META-INF/resources/images/empty-plant.png
