+++ appsec-kit-demo-v24-spring/src/main/resources/application.properties
