+++ appsec-kit-demo-v24-spring/src/main/resources/vaadin-featureflags.properties
