+++ pom.xml
@@ -139,7 +139,7 @@
                 <module>appsec-kit-demo-v8</module>
                 <module>appsec-kit-demo-v8-spring</module>
                 <module>appsec-kit-demo-v8-cdi</module>
-                <module>appsec-kit-demo-v24</module>
+                <module>appsec-kit-demo-v24-spring</module>
             </modules>
         </profile>
     </profiles>
